Rename utils module to models.
Rename the file to have a better name for this module.
